test/threads: fix jl_new_task slot test on 32-bit (Int64 -> native Int)
CI basic-lts-x86 (32-bit) failed: the jl_new_task slot-coherence test (R5-4)
hand-builds the `_foreigncall_` ABI with the ssize arg-type hardcoded as
`Val{Int64}()`, but the value `zl(0)` is the native `Int` (Int32 on 32-bit),
so the ccall errored `TypeError: in ccall argument 3, expected Int64, got
Int32`. Use `Val{Int}()` so the arg-type tracks the platform's native Int and
matches the value on both 32- and 64-bit (identical to before on x64).
Verified: rules/threads 14/14 on x64; the x86 ccall TypeError is resolved.

+++ test/rules/threads.jl
@@ -19,13 +19,15 @@ end
     # (dual_type(Task) === TaskTangent). The Task's V is never consumed in simple forward AD, so
     # this is checked at the frule boundary. `TaskTangent` is width-invariant, so width 1 suffices.
     # Args mirror the normalized `_foreigncall_(Val(:jl_new_task), RT, AT, nreq, cc, f, cf, ssize)`.
+    # `ssize` is a native `Int` (Int32 on 32-bit), so the arg-type must be `Int`, not a hardcoded
+    # `Int64` — the latter mismatches the `Int32` value on 32-bit and ccall-TypeErrors.
     @testset "jl_new_task slot coherence" begin
         zl(v) = Mooncake.zero_lifted(Val(1), v)
         r = Mooncake.frule!!(
             zl(Mooncake._foreigncall_),
             zl(Val(:jl_new_task)),
             zl(Val{Ref{Task}}()),
-            zl((Val{Any}(), Val{Any}(), Val{Int64}())),
+            zl((Val{Any}(), Val{Any}(), Val{Int}())),
             zl(Val{0}()),
             zl(Val{:ccall}()),
             zl(() -> nothing),
